Обновление способа получения файла конфигурации

+++ proxyServer/src/modules/admin/admin.controller.ts
@@ -1,23 +1,19 @@
 import { join } from "path";
-
-import { omit } from "ramda";
-
-import { Controller, Get, Post } from "@nestjs/common";
-
+import { Controller, Get } from "@nestjs/common";
 import { CacheService } from "services/cache.service";
-
-import { ProxyService } from "services/proxy.service";
-
-import prepareUserProfileToFront from "modules/users/formatters/prepareUserProfileToFront";
 
 @Controller("api")
 export class AdminController {
-  constructor(private cacheService: CacheService, private proxyService: ProxyService) {}
+  constructor(private cacheService: CacheService) {}
 
   @Get("admin/config")
   async getConfig(): Promise<string> {
-    this.cacheService.removeConfigCache("serverConfig/main-config.js");
-    this.cacheService.removeConfigCache("serverConfig/pages");
-    return JSON.stringify(require(join(process.cwd(), "src/assets/serverConfig/main-config.js")));
+    if (process.env.NODE_ENV === "development") {
+      this.cacheService.removeConfigCache("serverConfig/main-config.js");
+      this.cacheService.removeConfigCache("serverConfig/pages");
+      return JSON.stringify(require(join(process.cwd(), "src/assets/serverConfig/main-config.js")));
+    }
+
+    return JSON.stringify(await import("assets/serverConfig/main-config.js"));
   }
 }
